requirements.txt
lalsuite
gwpy
ltfatpy
astropy
healpy
versioneer
